+++ database/migrations/2024_08_30_082202_create_allocations_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('allocations', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('payment_id')->constrained()->onDelete('cascade');
-            $table->foreignId('policy_id')->constrained()->onDelete('cascade');
-            $table->decimal('allocation_amount', 15, 2);
-            $table->date('allocation_date');
-            $table->timestamps();
-        });
+        if (!\Schema::hasTable('allocations')) {
+            Schema::create('allocations', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('payment_id')->constrained()->onDelete('cascade');
+                $table->foreignId('policy_id')->constrained()->onDelete('cascade');
+                $table->decimal('allocation_amount', 15, 2);
+                $table->date('allocation_date');
+                $table->timestamps();
+            });
+        }
     }
     
     public function down()
